Move cli parser to params
