+++ GPTifier/CMakeLists.txt
@@ -50,7 +50,6 @@ include_directories(
 )
 add_executable(
   "gpt"
-  "src/cli.cpp"
   "src/help_messages.cpp"
   "src/main.cpp"
   "src/params.cpp"
