--- GPTifier/include/cli.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace cli
-{
-void parse_cli(const int argc, char **argv);
-}
